Fix occurrence of “Wrong type argument: stringp” error
It appears that the ‘feed_id’ is at least now a number rather than a string. I
don't know if it was different for older versions of Tiny Tiny RSS, but to
maintain the same functionality I'll assume that it can be either a number or a
string.

I've renamed ‘avandu--order-feed’ to ‘avandu-feed<’ because that more closely
matches the intention of the function.